requirements.txt
scanpy>=1.9.1
torch==1.12.1
torchvision==0.13.1
torchaudio==0.12.1
scikit-learn>=1.1.3
scvi-tools==0.19.0
scvi-colab
scip>=1.9.3
episcanpy==0.3.2
seaborn>=0.11.2
matplotlib>=3.6.2
pot==0.9.0
leidenalg
pybedtools
adjusttext
jupyter